This is a huge commit. I am so sorry. Major overhaul regarding the format of the repo. In particular:
We now have a configs folder for configuration files / functions.
We now have a models folder to store our database model.
We are now using a lot of modules from npm. Specifically, we started
using mongoose, pwd, passport, passport-local, on top of Nunjucks and
Express. Moreover, there are a lot more require calls overall, just due
to the nature of how I reformatted the repository.

The package.json should be set up properly, but it will not take into
account that this new branch requires mongoDB now. From here, I'll have
to write some basic scripts to get mongoDB up and running fairly easily
regardless of platform. I'll probably put those in a scripts/ folder. As
a last note on MongoDB, it seems desirable to run it on port 29999 for
the time being, since it is easier to set up a separate instance on a
non-default port in Debian. It's trivial to change for the user (edit
config/config.js) so I won't worry about it too much.